DI renamed to IoC
