+++ src/PrimitiveObsession/IoC/ConfigModule.cs
@@ -2,7 +2,7 @@
 using Autofac;
 using PrimitiveObsession.Config;
 
-namespace PrimitiveObsession.DI
+namespace PrimitiveObsession.IoC
 {
     public class ConfigModule : Module
     {
